refactor: fix things here and there
