+++ src/tck/include/key/KeyService.h
@@ -34,6 +34,31 @@ namespace Hedera::TCK::KeyService
 class GenerateKeyParams;
 
 /**
+ * Enumeration of the possible types of Keys the SDK server can generate.
+ */
+enum class KeyType : unsigned char
+{
+  ED25519_PRIVATE_KEY_TYPE,
+  ED25519_PUBLIC_KEY_TYPE,
+  ECDSA_SECP256k1_PRIVATE_KEY_TYPE,
+  ECDSA_SECP256k1_PUBLIC_KEY_TYPE,
+  LIST_KEY_TYPE,
+  THRESHOLD_KEY_TYPE,
+  EVM_ADDRESS_KEY_TYPE,
+  KEY_TYPE_SIZE [[maybe_unused]]
+};
+
+/**
+ * Map of KeyTypes to their string representation.
+ */
+extern const std::unordered_map<std::string, KeyType> gStringToKeyType;
+
+/**
+ * Map of KeyType string representations to KeyType.
+ */
+extern const std::unordered_map<KeyType, std::string> gKeyTypeToString;
+
+/**
  * Generate a Key.
  *
  * @param params The parameters to use to generate a key.
